+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph/structured_document_tag.rb
@@ -1,9 +1,12 @@
 require_relative 'sdt/sdt_content'
+require_relative 'sdt/sdt_properties'
 module OoxmlParser
   # Class for parsing `w:std`
   class StructuredDocumentTag < OOXMLDocumentObject
     # @return [SDTContent] content of sdt
     attr_reader :sdt_content
+    # @return [SDTProperties] properties of sdt
+    attr_reader :properties
 
     # Parse StructuredDocumentTag object
     # @param node [Nokogiri::XML:Element] node to parse
@@ -13,6 +16,8 @@ module OoxmlParser
         case node_child.name
         when 'sdtContent'
           @sdt_content = SDTContent.new(parent: self).parse(node_child)
+        when 'sdtPr'
+          @properties = SDTProperties.new(parent: self).parse(node_child)
         end
       end
       self